remove redudant parsers
