+++ packages/client/components/MeetingSeriesManagementRoot.tsx
@@ -1,4 +1,5 @@
 import MeetingSeriesId from 'parabol-client/shared/gqlIds/MeetingSeriesId'
+import {parseMeetingSeriesIdFromSlug} from 'parabol-client/shared/meetingSeriesSlug'
 import {Suspense} from 'react'
 import {Navigate, useParams} from 'react-router'
 import meetingSeriesManagementPageQuery, {
@@ -9,11 +10,8 @@ import MeetingSeriesManagementPage from './MeetingSeriesManagementPage'
 
 const MeetingSeriesManagementRoot = () => {
   const {meetingSeriesId: routeParam} = useParams()
-  const meetingSeriesId = routeParam
-    ? /^\d+$/.test(routeParam)
-      ? MeetingSeriesId.join(Number(routeParam))
-      : routeParam
-    : ''
+  const rawId = routeParam ? parseMeetingSeriesIdFromSlug(routeParam) : null
+  const meetingSeriesId = rawId != null ? MeetingSeriesId.join(rawId) : ''
   const queryRef = useQueryLoaderNow<MeetingSeriesManagementPageQuery>(
     meetingSeriesManagementPageQuery,
     {meetingSeriesId}
